Refactor preprocessing step.
