+++ src/Conversion/KrnlToAffine/CMakeLists.txt
@@ -2,6 +2,7 @@
 
 add_onnx_mlir_library(OMKrnlToAffine
   KrnlToAffine.cpp
+  KrnlLoopsLowering.cpp
 
   LINK_LIBS PUBLIC
   OMSpecializedKernelOpInterface
